This change addresses the feature request #61.
$smtgTagsProperties now allows callables among property names.

The callable is called and its return(s) are added to the tag values,
as if they were property values.

The callable is passed ( OutputPage $outputPage ), which allows access
to the current page metadata, like its title, canonical URL
or incoming redirects.

The messages are also fully initialised by the point the callable
is invoked.

Example:
$smtgTagsProperties = [
	'keywords' => [ 'Has keyword', function( OutputPage $outputPage ) {
		// Redirects to the page are collected as keywords.
		$redirects = $outputPage->getContext()->getTitle()->getRedirectsHere();
		$redirect_titles = [];
		foreach ( $redirects as $redirect ) {
			$redirect_titles[] = $redirect->getText();
		}
		return implode( ', ', array_unique( $redirect_titles ));
	} ]
];

Bug: https://github.com/SemanticMediaWiki/SemanticMetaTags/issues/61